Internal server no longer needs the data model endpoint, so move the code for the data model endpoint from shared python to external server.
