+++ components/external_server/src/routes/datamodel.py
@@ -3,10 +3,17 @@
 import bottle
 from pymongo.database import Database
 
-from shared.routes import get_data_model as get_data_model_implementation
+from shared.database.datamodels import latest_datamodel
+from shared.utils.functions import md5_hash, report_date_time
 
 
 @bottle.get("/api/v3/datamodel", authentication_required=False)
 def get_data_model(database: Database):
     """Return the data model."""
-    return get_data_model_implementation(database)  # pragma: no cover
+    data_model = latest_datamodel(database, report_date_time())
+    if data_model:
+        md5 = md5_hash(data_model["timestamp"])
+        if "W/" + md5 == bottle.request.headers.get("If-None-Match"):  # pylint: disable=no-member
+            bottle.abort(304)  # Data model unchanged
+        bottle.response.set_header("ETag", md5)
+    return data_model
