+++ components/internal_server/src/routes/__init__.py
@@ -1,6 +1,5 @@
 """Internal routes."""
 
-from .datamodel import get_data_model
 from .health import get_health
 from .measurement import post_measurement
 from .metric import get_metrics
