--- components/internal_server/src/routes/datamodel.py
@@ -1,12 +1,0 @@
-"""Data model routes."""
-
-import bottle
-from pymongo.database import Database
-
-from shared.routes import get_data_model as get_data_model_implementation
-
-
-@bottle.get("/api/datamodel")
-def get_data_model(database: Database):
-    """Return the data model."""
-    return get_data_model_implementation(database)  # pragma: no cover
