--- components/shared_python/src/shared/routes/__init__.py
@@ -1,3 +1,0 @@
-"""Routes."""
-
-from .datamodel import get_data_model
